Change lightcone map send size limit to double to allow values <1MB

+++ src/lightcone.h
@@ -81,7 +81,7 @@ struct lightcone_props {
   int hdf5_chunk_size;
 
   /* Maximum amount of data (in megabytes) to send from any one rank when updating healpix maps */
-  int max_map_update_send_size_mb;
+  double max_map_update_send_size_mb;
 
   /*! Whether to apply lossy compression */
   int particles_lossy_compression;
